RI2024G05-180
Changes requested

+++ app/src/main/java/com/example/riseandroid/network/auth0/Auth0DTO.kt
@@ -5,9 +5,8 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class AuthorizeResponse(
     val access_token: String,
-    val expires_in: Int,
+    val expires_in: Long,
     val token_type: String,
     val refresh_token: String?
 ) {
-
 }
